fix cast issue

+++ src/db/events-db.cpp
@@ -238,7 +238,7 @@ EventsDb::EventsDb () : AbstractDb(*new EventsDbPrivate) {}
 		isNull = false;
 
 		try {
-			return message.get<T>(index);
+			return message.get<T>((size_t)index);
 		} catch (const exception &) {
 			isNull = true;
 		}
